Initial version, does not support related resources with many
